django_web_server/home_server/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
